+++ src/main/java/net/llamaslayers/infiniteworld/OctreeGenerator.java
@@ -4,5 +4,5 @@ package net.llamaslayers.infiniteworld;
  * @author Nightgunner5
  */
 public interface OctreeGenerator {
-	public Material getMaterial(double x, double y, double z);
+	public Material getMaterial(float x, float y, float z);
 }
